backend/requirements.txt
asgiref-3.7.2
django-4.2.6
sqlparse-0.4.4
typing-extensions-4.8.0
